+++ doc/source/changelog.md
@@ -13,6 +13,8 @@ Dev 1.4.2
 
 Primarily developped with Godot 4.4.1+
 
+- `VoxelGeneratorGraph`: implemented constant reduction, which slightly optimizes graphs running on CPU if they contain constant branches
+
 - Fixes
     - `VoxelBlockyTypeLibrary`: fixed crash when setting `types` to empty array
     - `VoxelGeneratorGraph`: Editor: fixed error sometimes printing after closing the graph editor